Fractal made up of crosses using C#
Using recursion, I built a fractal out of crosses. After creating it, I
found out that the design I used to build it was already a famous
fractal. However since the code is 100% mine, this is just my take on
the famous fractal. i am especially proud of this project because of how
beautiful it looks and frankly because of how challenging it was to
build. After long hours of testing and debugging, the feeling I got when
the fractal worked is undescribable.